pyproject.toml
[tool.poetry]
name = "vajeh-pipeline-docker"
version = "0.1.0"
description = "docker container for common pipeline build and deploy"
authors = ["Jalal Hosseini <@artronics>"]
license = "MIT"
readme = "README.md"
packages = [{include = "vajeh_pipeline_docker"}]

[tool.poetry.group.dev.dependencies]
invoke = "^1.7.1"

[virtualenvs]
create = true
in-project = true
path = ".venv"

[tool.poetry.dependencies]
python = "^3.9"
toml = "^0.10.2"
semver = "^2.13.0"
GitPython = "^3.1.27"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
